binfile: Use gsize for the address...
Use gsize for the address and glib pointer cast to make
the compiler happy about what we do.

+++ src/binfile.h
@@ -35,7 +35,7 @@ typedef struct BinSymbol BinSymbol;
 BinFile *        bin_file_new           (const char      *filename);
 void             bin_file_free          (BinFile         *bin_file);
 const BinSymbol *bin_file_lookup_symbol (BinFile         *bin_file,
-					 gulong           address);
+					 gsize           address);
 gboolean         bin_file_check_inode   (BinFile         *bin_file,
 					 ino_t		  inode);
 const char *     bin_symbol_get_name    (BinFile         *bin_file,
